added pausable token

+++ contracts/erc20/ERC20Standard.sol
@@ -66,7 +66,7 @@ contract ERC20Standard {
         return _transfer(msg.sender, _to, _value);
     }
 
-    function transferFrom(address _from, address _to, uint _value) public returns (bool success) {
+    function transferFrom(address _from, address _to, uint256 _value) public returns (bool success) {
         uint256 allowed = dataStorage.getAllowance(_from, msg.sender);
         require(allowed >= _value);
 
@@ -76,7 +76,7 @@ contract ERC20Standard {
         return _transfer(_from, _to, _value);
     }
 
-    function approve(address _spender, uint _value) public returns (bool success) {
+    function approve(address _spender, uint256 _value) public returns (bool success) {
         if (dataStorage.setAllowance(msg.sender, _spender, _value)) {
             Approval(msg.sender, _spender, _value);
             return true;
